Fix: Fixed error once playing audio and then returning home. Added 'react-audio-player'
