+++ package.json
@@ -19,6 +19,7 @@
     "lucide-react": "^0.475.0",
     "next": "15.1.7",
     "react": "^19.0.0",
+    "react-audio-player": "^0.17.0",
     "react-dom": "^19.0.0",
     "react-pdf": "^9.2.1",
     "tailwind-merge": "^3.0.1",
