+++ src/components/custom/Dashboard.tsx
@@ -2,13 +2,10 @@
 
 import { SignedIn, SignedOut, SignInButton, SignOutButton, useUser } from "@clerk/nextjs";
 import { Button } from "../ui/button";
-import { useEffect, useMemo, useRef, useState } from "react";
-import { Document, Page, pdfjs } from 'react-pdf';
-import 'react-pdf/dist/esm/Page/TextLayer.css'; // Import the TextLayer CSS
-import 'react-pdf/dist/esm/Page/AnnotationLayer.css'; // Import AnnotationLayer CSS, if needed
-import AudioPlayer from "./Audioplayer";
+import { useEffect, useRef, useState } from "react";
+import { pdfjs } from 'react-pdf';
 import { motion } from "framer-motion";
-import { ChevronLeft, ChevronRight, Download, OctagonX, Trash2, Shuffle } from "lucide-react";
+import { Download, OctagonX, Trash2, Shuffle } from "lucide-react";
 import {
   Alert,
   AlertDescription,
@@ -24,8 +21,8 @@ import {
   AlertDialogHeader,
   AlertDialogTitle,
   AlertDialogTrigger,
-} from "../ui/alert-dialog"
-import { Separator } from "../ui/separator";
+} from "../ui/alert-dialog";
+import PdfView from "./PdfView";
 
 pdfjs.GlobalWorkerOptions.workerSrc = new URL(
   'pdfjs-dist/build/pdf.worker.min.mjs',
@@ -189,24 +186,6 @@ export default function Dashboard({ uploadPdf, uploadPdfMetadata, listPdfs, conv
     };
 
 
-    function onDocumentLoadSuccess({ numPages }: { numPages: number }): void {
-        setNumPages(numPages);
-        setPageNumber(1);
-    }
-
-    function changePage(offset: number) {
-        // setPageRendering(true);
-        setPageNumber(prevPageNumber => (prevPageNumber || 1) + offset);
-    }
-
-    function previousPage() {
-        changePage(-1);
-    }
-
-    function nextPage() {
-        changePage(1);
-    }
-
 
     const handleDragEnter = (e: React.DragEvent<HTMLDivElement>) => {
         e.preventDefault()
@@ -295,31 +274,6 @@ export default function Dashboard({ uploadPdf, uploadPdfMetadata, listPdfs, conv
             savePdf(file);
         }
     }
-
-    const MemoizedPage = useMemo(() => {
-        return (
-            <div
-                style={{
-                    width: '100%',
-                    height: '100%',
-                    display: 'flex',
-                    justifyContent: 'center',
-                    objectFit: 'contain',
-                }}
-            >
-                <Page 
-                    pageNumber={pageNumber}
-                    // onRenderSuccess={() => setPageRendering(false)}
-                    // onRenderError={() => setPageRendering(false)}
-                    renderTextLayer={true}
-                    renderAnnotationLayer={true}
-                    className="w-full h-full"
-                    scale={0.7}
-                    loading={<div className="flex justify-center p-10"><div className="animate-pulse">Loading page...</div></div>}
-                />
-            </div>
-        );
-    }, [pageNumber]);
 
 
     const deleteOnClick = async (id: string) => {
@@ -509,75 +463,7 @@ export default function Dashboard({ uploadPdf, uploadPdfMetadata, listPdfs, conv
                     </div>
                 </div>
             ) : (
-                <div className="bg-zinc-900 rounded-2xl overflow-hidden">
-                    <div className="flex flex-col gap-2 p-6">
-                        <div className="flex justify-between items-center">
-                            <Button 
-                                variant="ghost"
-                                className="hover:bg-zinc-800 hover:text-white" onClick={() => {
-                                    setPdfView(false);
-                                }}
-                            >Home</Button>
-                            <h2 className="text-xl font-medium">{pdfName}</h2>
-                        </div>
-                        <div className="flex justify-between items-center mb-6">
-                            <Button
-                                variant="ghost"
-                                className="hover:bg-zinc-800"
-                                disabled={(pageNumber || 1) <= 1}
-                                onClick={previousPage}
-                            >
-                                <ChevronLeft className="h-6 w-6" />
-                            </Button>
-                            <span className="text-sm text-zinc-400">
-                                Page {pageNumber || (numPages ? 1 : '--')} of {numPages || '--'}
-                            </span>
-                            <Button
-                                variant="ghost"
-                                className="hover:bg-zinc-800"
-                                disabled={(pageNumber || 0) >= (numPages || 0)}
-                                onClick={nextPage}
-                            >
-                                <ChevronRight className="h-6 w-6" />
-                            </Button>
-                        </div>
-                        <div className="flex flex-row w-full gap-4 ">
-                            
-                            <div className="flex flex-col gap-2 w-1/2">
-                                <h2 className="text-lg font-semibold">Pdf Viewer</h2>   
-                                <Separator className="bg-gray-700"/>
-                                <div className="pdf-container h-[calc(100vh-420px)] min-h-[400px] overflow-y-auto w-full flex justify-center">
-                                    <Document
-                                        file={pdfFile}
-                                        onLoadSuccess={onDocumentLoadSuccess}
-                                    >
-                                        {MemoizedPage}
-                                    </Document>
-                                </div>
-
-                                <div className="flex items-center justify-between gap-1">
-                                    <div className="flex items-center space-x-4">
-                                        <AudioPlayer 
-                                            text={pagesText[(pageNumber || 1) - 1] || ""}
-                                            convertTextToSpeech={convertTextToSpeech}
-                                        />
-                                    </div>
-
-                                </div>
-                            </div>
-                            <div className="flex flex-col gap-2 w-1/2">
-                                    <h2 className="text-lg font-semibold">Current Page</h2>
-                                    <Separator className="bg-gray-700" />
-                                    <div className="flex flex-row justify-center">
-                                        <div className="flex flex-col text-center h-[calc(100vh-422px)] overflow-y-auto text-md w-10/12">
-                                            {pagesText[(pageNumber || 1)-1] || "No text found on this page"}
-                                        </div>
-                                    </div>
-                            </div>
-                        </div>
-
-                    </div>
-                </div>
+                <PdfView setPdfView={setPdfView} pdfName={pdfName} numPages={numPages || 1} pdfFile={pdfFile} pagesText={pagesText} convertTextToSpeech={convertTextToSpeech} setNumPages={setNumPages} pageNumber={pageNumber || 1} setPageNumber={setPageNumber} />
             )}
         </div>
   </div>);
